+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+X.Y.Z Release notes
+=============================================================
+### Breaking changes
+* None.
+
+### Enhancements
+* None.
+
+### Bug fixes
+* None.
+
+### Internal
+* Added property `Realm._numberOfVersions()` to help internal debugging.
+
 2.2.6 Release notes (2018-1-26)
 =============================================================
 ### Breaking changes